fix README and config.yaml
In `config.yaml`, a few things were specified incorrectly: fix.

The README was partially incorrect or contained information not relevant to this
repo. In general, accurate long READMEs with details are good, but short accurate README
files are better than long inaccurate ones. So I got rid of stuff that is not relevant
or incorrect for this repo.